requirements.txt
mysql-connector-python==8.3.0
aiomqtt==2.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.29
pandas==2.2.1